Applying changes from code review
