+++ lib/nexpose.rb
@@ -65,6 +65,7 @@ require 'nexpose/alert'
 require 'nexpose/ajax'
 require 'nexpose/api_request'
 require 'nexpose/asset'
+require 'nexpose/blackout'
 require 'nexpose/common'
 require 'nexpose/console'
 require 'nexpose/credential'
